\name{fdth-internal}
\title{Internal fdth functions}
\alias{make.fdt.cat}
\alias{make.fdt.cat.simple}
\alias{make.fdt.cat.multiple}
\alias{make.fdt.format.classes}
\alias{make.fdt.multiple}
\alias{make.fdt}
\alias{make.fdt.simple}
\description{Internal fdth functions.}
\details{These are not to be called by the user and are undocumented.}
\keyword{internal}
