% ***Enio: eu adaptei para para as categóricas, necessário correção***
\name{fdt}
\alias{fdt}
\alias{fdt.default}
\alias{fdt.data.frame}
\alias{fdt.matrix}

\title{
  Frequency distribution table for numerical
}

\description{
  A S3 set of methods to easily perform frequency distribution table (\samp{fdt}) from
  \code{vector}, \code{data.frame} and \code{matrix} objects.
}

\usage{
## S3 generic
fdt(x, \dots)

## S3 methods
\method{fdt}{default}(x,
    k,
    start,
    end,
    h,
    breaks=c('Sturges', 'Scott', 'FD'),
    right=FALSE, \dots)

\method{fdt}{data.frame}(x,
    k,
    by,
    breaks=c('Sturges', 'Scott', 'FD'),
    right=FALSE, \dots)

\method{fdt}{matrix}(x,
    k,
    breaks=c('Sturges', 'Scott', 'FD'),
    right=FALSE, \dots)
}

\arguments{
  \item{x}{A \code{vector}, \code{data.frame} or \code{matrix} object.
    If \samp{x} is \code{data.frame} or \code{matrix} it must contain at least 
    one numeric column (\code{fdt}) or character/factor (\code{fdt.cat}).}
  \item{k}{Number of class intervals.}
  \item{start}{Left endpoint of the first class interval.}
  \item{end}{Right endpoint of the last class interval.}
  \item{h}{Class interval width.}
  \item{by}{Categorical variable used for grouping each numeric variable,
    useful only on \code{data.frame}.}
  \item{breaks}{Method used to determine the number of interval classes, 
    c(\dQuote{Sturges}, \dQuote{Scott}, \dQuote{FD}).}
  \item{right}{Right endpoints open (default = \code{FALSE}).}
  \item{\dots}{Potencial further arguments (required by generic).}
}

\details{
  The simplest way to run \samp{fdt} is done by supplying only the \samp{x}
  object, for example: \code{nm <- fdt(x)}. In this case all necessary
  default values (\samp{breaks} and \samp{right}) (\dQuote{Sturges} and \code{FALSE}
  respectively) will be used.

  It can be provided also:
  \itemize{
    \item \samp{x} and \samp{k} (number of class intervals);
    \item \samp{x}, \samp{start} (left endpoint of the first class interval) and
    \samp{end} (right endpoint of the last class interval); or
    \item \samp{x}, \samp{start}, \samp{end} and \samp{h} (class interval width).
  }
  These options make the \samp{fdt} very easy and flexible.

  The \samp{fdt} object stores information to be used by methods \code{summary},
  \code{print} and \code{plot}. The result of plot is a histogram.
  The methods \code{summary}, \code{print} and \code{plot} provide a reasonable
  set of parameters to format and plot the \samp{fdt} object in a pretty
  (and publishable) way.
}

\value{
  For \code{fdt} the method \code{fdt.default} returns a list of class \code{fdt.default} with the slots:
  \item{\samp{table}}{A \code{data.frame} storing the \samp{fdt};}
  \item{\samp{breaks}}{A \code{vector} of length 4 storing \samp{start}, \samp{end},
    \samp{h} and \samp{right} of the \samp{fdt} generated by this method;}
  \item{\samp{data}}{A vector of the data \samp{x} provided.}

  The methods \code{fdt.data.frame} and \code{fdt.matrix}
  return a list of class \code{fdt.multiple}.
  This \code{list} has one slot for each numeric (\code{fdt})
  variable of the \samp{x} provided. Each slot, corresponding to each numeric
  variable, stores the same slots of the \code{fdt.default} described above.
}

\author{
  José Cláudio Faria \cr
  Enio G. Jelihovschi
}

\seealso{
  \code{\link[graphics]{hist}} provided by \pkg{graphics};
  \code{\link[base]{table}}, \code{\link[base]{cut}} both provided by \pkg{base}
  and \code{\link[Hmisc]{hist.data.frame}} provided by \pkg{Hmisc} package.
}

\examples{
library(fdth)

##========
## Vector
##========
x <- rnorm(n=1e3,
           mean=5,
           sd=1)

# x
(fdt <- fdt(x))

# x, alternative breaks
(fdt <- fdt(x,
            breaks='Scott'))

# x, k
(fdt <- fdt(x,
            k=10))

# x, star, end
range(x)

(fdt <- fdt(x,
            start=floor(min(x)),
            end=floor(max(x) + 1)))

# x, start, end, h
(fdt <- fdt(x,
            start=floor(min(x)),
            end=floor(max(x) + 1),
            h=1))

# Effect of right
x <- rep(1:3, 3); sort(x)

(fdt <- fdt(x,
            start=1,
            end=4,
            h=1))

(fdt <- fdt(x,
            start=0,
            end=3,
            h=1,
            right=TRUE))

##================================================
## Data.frame: multivariated with two categorical
##================================================
mdf <- data.frame(c1=sample(LETTERS[1:3], 1e2, TRUE),
                  c2=as.factor(sample(1:10, 1e2, TRUE)),
                  n1=c(NA, NA, rnorm(96, 10, 1), NA, NA),
                  n2=rnorm(100, 60, 4),
                  n3=rnorm(100, 50, 4))

head(mdf)

(fdt <- fdt(mdf))

# By factor!
(fdt <- fdt(mdf,
            k=5,
            by='c1'))

# choose FD criteria               
(fdt <- fdt(mdf,
            breaks='FD',
            by='c1'))

(fdt <- fdt(mdf,
            k=5,
            by='c2'))

(fdt <- fdt(iris,
            k=10))

(fdt <- fdt(iris,
            k=5,
            by='Species'))

#=========================
# Matrices: multivariated
#=========================
(fdt <-fdt(state.x77))
}

\keyword{fdt}
\keyword{fdt.cat}
\keyword{frequency}
\keyword{distribution}
\keyword{table}
