\name{fdth-internal}
\title{Internal fdth functions}
\alias{make.fdt_cat.simple}
\alias{make.fdt_cat.multiple}
\alias{make.fdt.format.classes}
\alias{make.fdt.multiple}
\alias{make.fdt.simple}
\description{Internal fdth functions.}
\details{These are not to be called by the user and are undocumented.}
\keyword{internal}
