\name{make.fdt}
\alias{make.fdt}
\alias{make.fdt_cat}

\title{
  Frequency distribution table for continuous and categorical variables
}

\description{
  Makes a full fdt from a minimal set of information. \cr
  Useful to reproduce (when the real data vector is not known) a previous fdt.
}

\usage{
make.fdt(f,
         start,
         end,
         right=FALSE)

make.fdt_cat(f,
             categories=NULL,
             sort=TRUE,
             decreasing=TRUE)
}

\arguments{
  \item{f}{A numeric \code{vector} object of frequency.}
  \item{start}{The left value of the interval of the first class.}
  \item{end}{The last value of the interval of the last class.}
  \item{right}{Intervals right open (default = \code{FALSE}).}
  \item{categories}{...}
  \item{sort}{...}
  \item{decreasing}{...}
}

\details{ Given the starting and ending values of the continuous variable table 
 or the levels of the categorical variable plus the number of intervals and the 
 absolute frequency values the functions \code{make.fdt} and \code{make.fdt_cat}
 reconstruct whole fdt or fdt_cat table.   
}

\value{
  The function \code{make.fdt} returns a list with the slots:
  \item{table}{A \code{data.frame} storing the \samp{fdt}.}
  \item{breaks}{A \code{vector} of length 4 storing \samp{start}, \samp{end},
    \samp{h} and \samp{right} of the \samp{fdt} generated by this method.}
  The function \code{make.fdt_cat} returns a list whith the slots:
  \item{Category}{The levels of the categorical variable.}
  \item{f}{Absolute frequency, \code{numeric}}
  \item{rf}{Relative frequency, \code{numeric}}
  \item{rf(\%)}{Relative frequency in percentages, \code{numeric}}
  \item{cf}{Cumulative frequency; \code{numeric}}
  \item{cf(\%)}{Cumulative frequency in percentages, \code{numeric}}
}

\seealso{
  \code{\link[base]{table}} and \code{\link[base]{cut}} provided by \pkg{base} package.
}

\author{
  José Cláudio Faria \cr
  Enio G. Jelihovschi \cr
  Ivan B. Allaman
}


\examples{
## Numeric
## Making one reference fdt
set.seed(33)
x <- rnorm(1e3, 
           20, 
           2)

(tb.r <- fdt(x))

## Making a brand new
(tb.n  <- make.fdt(f=tb.r$table$f,
                   start=13.711,
                   end=27.229))   # Huumm ..., good, but ... Can it be better? 

summary(tb.n,
        format=TRUE,
        pattern='\%.3f')           # Is it nice now?


## Categorical
x <- sample(letters[1:5], 
            1e3, 
            rep=TRUE)

## Making one reference fdt
(tb.r <- fdt_cat(x))

## Making a brand new
(tb.n <- make.fdt_cat(f=tb.r$f,
                      categ=tb.r$Category))
}
